+++ aeron-client/src/main/c/collections/aeron_int64_to_ptr_hash_map.h
@@ -48,7 +48,7 @@ inline int aeron_int64_to_ptr_hash_map_init(
     size_t capacity = (size_t)aeron_find_next_power_of_two((int32_t)initial_capacity);
 
     map->load_factor = load_factor;
-    map->resize_threshold = (size_t)(load_factor * capacity);
+    map->resize_threshold = (size_t)(load_factor * (float)capacity);
     map->keys = NULL;
     map->values = NULL;
     map->capacity = capacity;
@@ -83,7 +83,7 @@ inline void aeron_int64_to_ptr_hash_map_delete(aeron_int64_to_ptr_hash_map_t *ma
 inline int aeron_int64_to_ptr_hash_map_rehash(aeron_int64_to_ptr_hash_map_t *map, size_t new_capacity)
 {
     size_t mask = new_capacity - 1;
-    map->resize_threshold = (size_t)(new_capacity * map->load_factor);
+    map->resize_threshold = (size_t)((float)new_capacity * map->load_factor);
 
     int64_t *tmp_keys;
     void **tmp_values;
